ISSUE #3062 model helper added with removeModelData method swagger doc added
